Now pub version #2 is the default.  Jason, it seems as if this fixes the republication crash you showed me a couple of days ago.
The problem was that the PUB_FILES RPC was generating a response that
was marginally too big, since it was sending a huge collection of
pub'ed files over IPC.  In pub version 2, this transfer is split
up so that files can be fetched individually.

This fix also makes use of libpub/rpcwindow.[Ch], which is a generic
class for making N RPCS with concurrency W, where N >> W.  This
might be really useful in other places where we roll it by hand now.


git-svn-id: svn://okws.org/trunk/okws2@1090 7d287422-6eea-da11-b677-00123f2a783a