Fix eggregious bug in XML parsing that was breaking e-mails being send by CBD.
The bug was triggered when a <name>...</name> field was broken over
a package boundary.  In that case, I was ignoring the data from packet 2 and
only keeping the data from packet 1.

This bug was more likely to show up in a bigger packet that has many fields
but small  payloads.  When string data was broken over the packet boundary,
things still worked.






git-svn-id: svn://okws.org/okws2/devel/1.2@3086 7d287422-6eea-da11-b677-00123f2a783a